+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 <details>
 <summary>View Changelog</summary>
 
+# 2.19.2
+- Fixed activated abilities not being interactable in Act 3
+- Fixed cards with costs above vanilla defaults not displaying
+- Added debug logs to AddCustomTribesToList (used to add custom Tribes to the list of obtainable Totem tops)
+
 # 2.19.1
 - Fixed API not retrieving pixel card costs above 5